Add labels to bot-created gitLab merge requests
GitLab doesn't seem to support matching 'bot users when filtering
merge requests, making it hard to find human merge requests if there
are a lot of 'bot requests open. GitLab does support filtering on
labels, so this change allows a comma-delimited list of labels (as in
the GitLab merge request API) to be added to each merge request
opened.

A new CLI option is added to the script. The labels default to a
single label "npg_automation". This is intended to be a general label
that may be applied to many repositories and forms of automation. It
is intended to indicate that an item is the _product_ of automation.